+++ routes/blog.route.ts
@@ -1,31 +1,10 @@
-import { Context, DefaultState, Next } from "koa";
+import { Context, DefaultState } from "koa";
 import Router from "@koa/router";
 
-import { getBlogIndexArray, parseMarkdown } from "../utils/parsers";
+import { getBlogIndex, getBlog } from "../controllers/blog.controller";
 
 const blogRouter = new Router<DefaultState, Context>({ prefix: "/blog" });
-blogRouter.get("/", async (ctx: Context, next: Next) => {
-  try {
-    const blogsArray = await getBlogIndexArray();
-    await ctx.render("blogindex", { blogs: blogsArray });
-  } catch (error) {
-    await next();
-  }
-});
-blogRouter.get("/:id", async (ctx: Context, next: Next) => {
-  const blogId = ctx.params.id;
-
-  try {
-    const parsedObj = await parseMarkdown(blogId);
-    await ctx.render("blogpost", {
-      body: parsedObj.body,
-      title: parsedObj.header.title,
-      date: parsedObj.header.date,
-      tags: parsedObj.header.tags,
-    });
-  } catch (error) {
-    await next();
-  }
-});
+blogRouter.get("/", getBlogIndex);
+blogRouter.get("/:id", getBlog);
 
 export default blogRouter;