+++ routes/series.route.ts
@@ -1,13 +1,10 @@
 import { Context, DefaultState, Next } from "koa";
 import Router from "@koa/router";
 
+import { getSeriesIndex, getSeries } from "../controllers/series.controller";
+
 const seriesRouter = new Router<DefaultState, Context>({ prefix: "/series" });
-seriesRouter.get("/", async (ctx: Context, next: Next) => {
-  try {
-    await ctx.render("series");
-  } catch (error) {
-    await next();
-  }
-});
+seriesRouter.get("/", getSeriesIndex);
+seriesRouter.get("/:series", getSeries);
 
 export default seriesRouter;