+++ CHANGELOG.md
@@ -7,6 +7,18 @@ Addition of markdown posts will not be versioned or require a changelog update.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.3-Beta.2] - 2021-07-15
+
+### Added
+
+- /controllers/blog.controller.ts
+- /controllers/series.controller.ts
+
+### Changed
+
+- series.ejs => seriesindex.ejs
+- moved handlers out of routes and into separate controllers
+
 ## [0.2.3-Beta.1] - 2021-07-15
 
 ### Added